Flattening Algorithm.
